requirements.txt
matplotlib==3.8.2
tqdm
scikit-image==0.22.0
webdataset==0.2.73
pandas==2.2.0
einops
ftfy
h5py
regex
kornia==0.7.1
h5py==3.10.0
open_clip_torch==2.24.0
transformers==4.37.2
torchmetrics==1.3.0.post0
diffusers==0.23.0
wandb
omegaconf==2.3.0
pytorch-lightning==2.0.1
sentence-transformers==2.5.1
evaluate==0.4.1
nltk==3.8.1
rouge_score==0.1.2
umap==0.1.1
huggingface-hub==0.25.2
clip-anytorch==2.6.0
dalle2-pytorch==1.15.6
timm==0.6.12
pycocotools

